修改了IP以及清除了一些警告

+++ utils/pts/pts_model.cpp
@@ -228,7 +228,7 @@ QVariant pts_model::get_std(const QModelIndex &index, int role) const
 
 void pts_model::init_row(unsigned row)
 {
-    assert (row < rowCount ());
+    assert (static_cast<int> (row) < rowCount ());
     auto col = get_header_col (this, "数量*频次");
     if (col > 0)
     {
